+++ packages/types/src/interfaces/system/definitions.ts
@@ -214,6 +214,10 @@ export default {
       index: 'u8',
       error: 'u8'
     },
+    DispatchErrorModuleU8a: {
+      index: 'u8',
+      error: '[u8; 4]'
+    },
     DispatchErrorTo198: {
       module: 'Option<u8>',
       error: 'u8'